Added color reading

+++ src/matrix.c
@@ -6,17 +6,17 @@
 /*   By: zfaria <zfaria@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/28 15:30:00 by zfaria            #+#    #+#             */
-/*   Updated: 2019/03/29 11:51:46 by zfaria           ###   ########.fr       */
+/*   Updated: 2019/03/30 13:24:03 by zfaria           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <math.h>
 #include <fdf.h>
 
-void	rotate_x(t_coor *point, float thetacos, float thetasin)
+void	rotate_x(t_coor *point, double thetacos, double thetasin)
 {
 	int		y;
-	float	z;
+	double	z;
 
 	if (!point)
 		return ;
@@ -26,10 +26,10 @@ void	rotate_x(t_coor *point, float thetacos, float thetasin)
 	point->z = -y * thetasin + z * thetacos;
 }
 
-void	rotate_y(t_coor *point, float thetacos, float thetasin)
+void	rotate_y(t_coor *point, double thetacos, double thetasin)
 {
 	int		x;
-	float	z;
+	double	z;
 
 	if (!point)
 		return ;
@@ -39,7 +39,7 @@ void	rotate_y(t_coor *point, float thetacos, float thetasin)
 	point->z = -x * thetasin + z * thetacos;
 }
 
-void	rotate_z(t_coor *point, float thetacos, float thetasin)
+void	rotate_z(t_coor *point, double thetacos, double thetasin)
 {
 	int x;
 	int y;
